r/storage_share: refactoring the migrations out
